e2e: rename testdata package to testscenario
