+++ test/e2e/testscenario/cudn/invalid-scenarios-mismatch-topo-conf.go
@@ -1,8 +1,8 @@
 package cudn
 
-import "github.com/ovn-org/ovn-kubernetes/test/e2e/testdata"
+import "github.com/ovn-org/ovn-kubernetes/test/e2e/testscenario"
 
-var MismatchTopologyConfig = []testdata.ValidateCRScenario{
+var MismatchTopologyConfig = []testscenario.ValidateCRScenario{
 	{
 		Description: "topology is localnet but topology config is layer2",
 		ExpectedErr: `spec.localnet is required when topology is Localnet and forbidden otherwise`,
