+++ test/e2e/testscenario/scenario.go
@@ -1,4 +1,4 @@
-package testdata
+package testscenario
 
 // ValidateCRScenario represent test scenario where a manifest is applied and failed with the expected error
 type ValidateCRScenario struct {
